+++ core/src/com/angian/kungfu/GameConstants.kt
@@ -6,6 +6,7 @@ object GameConstants {
     const val WORLD_HEIGHT = 800f
 
     const val FIGHTER_WALK_SPEED = 100f //px per sec
+    const val ANIM_FRAME_DURATION = 0.1f
 }